Linter {lintr}	R Documentation

Create a linter closure

Description

Create a linter closure

Usage

Linter(fun, name = linter_auto_name())

Arguments

fun
A function that takes a source file and returns lint objects.

name
Default name of the Linter. Lints produced by the linter will be labelled with name by default.

Value

The same function with its class set to 'linter'.

[Package lintr version 3.1.2 Index]