T_and_F_symbol_linter {lintr}	R Documentation

T and F symbol linter

Description

Avoid the symbols T and F, and use TRUE and FALSE instead.

Usage

T_and_F_symbol_linter()

Tags

best_practices, consistency, default, readability, robustness, style

See Also

linters for a complete list of linters available in lintr.

https://style.tidyverse.org/syntax.html#logical-vectors

Examples

# will produce lints
lint(
  text = "x <- T; y <- F",
  linters = T_and_F_symbol_linter()
)

lint(
  text = "T = 1.2; F = 2.4",
  linters = T_and_F_symbol_linter()
)

# okay
lint(
  text = "x <- c(TRUE, FALSE)",
  linters = T_and_F_symbol_linter()
)

lint(
  text = "t = 1.2; f = 2.4",
  linters = T_and_F_symbol_linter()
)

[Package lintr version 3.1.2 Index]