all_linters {lintr}	R Documentation

Create a linter configuration based on all available linters

Description

Create a linter configuration based on all available linters

Usage

all_linters(packages = "lintr", ...)

Arguments

packages
A character vector of packages to search for linters.

...
Arguments of elements to change. If unnamed, the argument is automatically named. If the named argument already exists in the list of linters, it is replaced by the new element. If it does not exist, it is added. If the value is NULL, the linter is removed.

See Also

linters_with_defaults for basing off lintr's set of default linters.

linters_with_tags for basing off tags attached to linters, possibly across multiple packages.

available_linters to get a data frame of available linters.

linters for a complete list of linters available in lintr.

Examples

names(all_linters())

[Package lintr version 3.1.2 Index]