best_practices_linters {lintr}	R Documentation

Best practices linters

Description

Linters checking the use of coding best practices, such as explicit typing of numeric constants.

Linters

The following linters are tagged with 'best_practices':

absolute_path_linter

any_duplicated_linter

any_is_na_linter

boolean_arithmetic_linter

class_equals_linter

commented_code_linter

condition_message_linter

conjunct_test_linter

cyclocomp_linter

empty_assignment_linter

expect_comparison_linter

expect_length_linter

expect_named_linter

expect_not_linter

expect_null_linter

expect_s3_class_linter

expect_s4_class_linter

expect_true_false_linter

expect_type_linter

extraction_operator_linter

fixed_regex_linter

for_loop_index_linter

function_argument_linter

function_return_linter

ifelse_censor_linter

implicit_assignment_linter

implicit_integer_linter

is_numeric_linter

length_levels_linter

lengths_linter

library_call_linter

literal_coercion_linter

nonportable_path_linter

outer_negation_linter

paste_linter

redundant_equals_linter

redundant_ifelse_linter

regex_subset_linter

routine_registration_linter

scalar_in_linter

seq_linter

sort_linter

system_file_linter

T_and_F_symbol_linter

undesirable_function_linter

undesirable_operator_linter

unnecessary_lambda_linter

unnecessary_nested_if_linter

unnecessary_placeholder_linter

unreachable_code_linter

unused_import_linter

vector_logic_linter

yoda_test_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]