configurable_linters {lintr}	R Documentation

Configurable linters

Description

Generic linters which support custom configuration to your needs.

Linters

The following linters are tagged with 'configurable':

absolute_path_linter

assignment_linter

backport_linter

brace_linter

commas_linter

conjunct_test_linter

cyclocomp_linter

duplicate_argument_linter

fixed_regex_linter

if_not_else_linter

implicit_assignment_linter

implicit_integer_linter

indentation_linter

infix_spaces_linter

library_call_linter

line_length_linter

missing_argument_linter

namespace_linter

nonportable_path_linter

object_length_linter

object_name_linter

object_usage_linter

paste_linter

pipe_consistency_linter

quotes_linter

redundant_ifelse_linter

semicolon_linter

string_boundary_linter

todo_comment_linter

trailing_whitespace_linter

undesirable_function_linter

undesirable_operator_linter

unnecessary_concatenation_linter

unused_import_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]