consistency_linters {lintr}	R Documentation

Consistency linters

Description

Linters checking enforcing a consistent alternative if there are multiple syntactically valid ways to write something.

Linters

The following linters are tagged with 'consistency':

assignment_linter

class_equals_linter

condition_message_linter

consecutive_assertion_linter

function_argument_linter

if_not_else_linter

implicit_integer_linter

inner_combine_linter

is_numeric_linter

keyword_quote_linter

length_levels_linter

literal_coercion_linter

numeric_leading_zero_linter

object_name_linter

paste_linter

quotes_linter

redundant_ifelse_linter

scalar_in_linter

seq_linter

system_file_linter

T_and_F_symbol_linter

whitespace_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]