default_linters {lintr}	R Documentation

Default linters

Description

List of default linters for lint(). Use linters_with_defaults() to customize it. Most of the default linters are based on the tidyverse style guide.

The set of default linters is as follows (any parameterized linters, e.g., line_length_linter use their default argument(s), see ⁠?<linter_name>⁠ for details):

Usage

default_linters

Format

An object of class list of length 25.

Linters

The following linters are tagged with 'default':

assignment_linter

brace_linter

commas_linter

commented_code_linter

cyclocomp_linter

equals_na_linter

function_left_parentheses_linter

indentation_linter

infix_spaces_linter

line_length_linter

object_length_linter

object_name_linter

object_usage_linter

paren_body_linter

pipe_continuation_linter

quotes_linter

semicolon_linter

seq_linter

spaces_inside_linter

spaces_left_parentheses_linter

T_and_F_symbol_linter

trailing_blank_lines_linter

trailing_whitespace_linter

vector_logic_linter

whitespace_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]