deprecated_linters {lintr}	R Documentation

Deprecated linters

Description

Linters that are deprecated and provided for backwards compatibility only. These linters will be excluded from linters_with_tags() by default.

Linters

The following linters are tagged with 'deprecated':

closed_curly_linter

consecutive_stopifnot_linter

no_tab_linter

open_curly_linter

paren_brace_linter

semicolon_terminator_linter

single_quotes_linter

unneeded_concatenation_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]