efficiency_linters {lintr}	R Documentation

Efficiency linters

Description

Linters highlighting code efficiency problems, such as unnecessary function calls.

Linters

The following linters are tagged with 'efficiency':

any_duplicated_linter

any_is_na_linter

boolean_arithmetic_linter

fixed_regex_linter

ifelse_censor_linter

inner_combine_linter

length_test_linter

lengths_linter

literal_coercion_linter

matrix_apply_linter

nested_ifelse_linter

outer_negation_linter

redundant_equals_linter

redundant_ifelse_linter

regex_subset_linter

routine_registration_linter

scalar_in_linter

seq_linter

sort_linter

string_boundary_linter

undesirable_function_linter

undesirable_operator_linter

unnecessary_concatenation_linter

unnecessary_lambda_linter

vector_logic_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]