exclude {lintr}	R Documentation

Exclude lines or files from linting

Description

Exclude lines or files from linting

Usage

exclude(lints, exclusions = settings$exclusions, ...)

Arguments

lints
that need to be filtered.

exclusions
manually specified exclusions

...
additional arguments passed to parse_exclusions

Details

Exclusions can be specified in three different ways.

single line in the source file. default: # nolint

line range in the source file. default: # nolint start, # nolint end

exclusions parameter, a named list of the files and lines to exclude, or just the filenames if you want to exclude the entire file.

[Package lintr version 1.0.3 Index]