executing_linters {lintr}	R Documentation

Code executing linters

Description

Linters that evaluate parts of the linted code, such as loading referenced packages. These linters should not be used with untrusted code, and may need dependencies of the linted package or project to be available in order to function correctly. For package authors, note that this includes loading the package itself, e.g. with pkgload::load_all() or installing and attaching the package.

Linters

The following linters are tagged with 'executing':

namespace_linter

object_length_linter

object_name_linter

object_usage_linter

unused_import_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]