lint_file {lintr}	R Documentation

Lint a given file

Description

Apply one or more linters to a file and return a list of lints found.

Usage

lint(filename, linters = NULL, cache = FALSE, ..., parse_settings = TRUE)

Arguments

filename
the given filename to lint.

linters
a list of linter functions to apply see linters for a full list of default and available linters.

cache
toggle caching of lint results, if passed a character vector uses that file as the cache.

...
additional arguments passed to exclude.

parse_settings
whether to try and parse the settings

[Package lintr version 1.0.3 Index]