lint_package {lintr}	R Documentation

Lint a package

Description

Apply one or more linters to all of the R files in a package.

Usage

lint_package(path = ".", relative_path = TRUE, ...)

Arguments

path
the path to the base directory of the package, if NULL, the base directory will be searched for by looking in the parent directories of the current directory.

relative_path
if TRUE, file paths are printed using their path relative to the package base directory. If FALSE, use the full absolute path.

...
additional arguments passed to lint

[Package lintr version 1.0.3 Index]