absolute_paths_linter {lintr}	R Documentation

linters

Description

Available linters

Usage

absolute_paths_linter(source_file)

assignment_linter(source_file)

closed_curly_linter(allow_single_line = FALSE)

commas_linter(source_file)

commented_code_linter(source_file)

infix_spaces_linter(source_file)

line_length_linter(length)

no_tab_linter(source_file)

object_usage_linter(source_file)

camel_case_linter(source_file)

snake_case_linter(source_file)

multiple_dots_linter(source_file)

object_length_linter(length = 20L)

open_curly_linter(allow_single_line = FALSE)

single_quotes_linter(source_file)

spaces_inside_linter(source_file)

spaces_left_parentheses_linter(source_file)

trailing_blank_lines_linter(source_file)

trailing_whitespace_linter(source_file)

Arguments

source_file
returned by get_source_expressions

allow_single_line
if true allow a open and closed curly pair on the same line.

length
the length cutoff to use for the given linter.

Functions

absolute_paths_linter: checks that no absolute paths are used.

assignment_linter: checks that '<-' is always used for assignment

closed_curly_linter: check that closed curly braces should always be on their own line unless they follow an else.

commas_linter: check that all commas are followed by spaces, but do not have spaces before them.

commented_code_linter: checks that there is no commented code outside roxygen blocks

infix_spaces_linter: check that all infix operators have spaces around them.

line_length_linter: check the line length of both comments and code is less than length.

no_tab_linter: check that only spaces are used, never tabs.

object_usage_linter: checks that closures have the proper usage using checkUsage. Note this runs eval on the code, so do not use with untrusted code.

camel_case_linter: check that objects are not in camelCase.

snake_case_linter: check that objects are not in snake_case.

multiple_dots_linter: check that objects do not have.multiple.dots.

object_length_linter: check that objects do are not very long.not have.multiple.dots.

open_curly_linter: check that opening curly braces are never on their own line and are always followed by a newline.

single_quotes_linter: checks that only single quotes are used to delimit string contestants.

spaces_inside_linter: check that parentheses and square brackets do not have spaces directly inside them.

spaces_left_parentheses_linter: check that all left parentheses have a space before them unless they are in a function call.

trailing_blank_lines_linter: check there are no trailing blank lines.

trailing_whitespace_linter: check there are no trailing whitespace characters.

[Package lintr version 1.0.3 Index]