lintr {lintr}	R Documentation

Lintr

Description

Checks adherence to a given style, syntax errors and possible semantic issues. Supports on the fly checking of R code edited with Emacs, Vim and Sublime Text.

See Also

lint, lint_package, linters

[Package lintr version 1.0.3 Index]