lintr-deprecated {lintr}	R Documentation

Deprecated functions in lintr

Description

These functions have been deprecated from lintr.

open_curly_linter() and closed_curly_linter() check that open and closed curly braces are on their own line unless they follow an else, a comma, or a closing bracket. Deprecated in favor of brace_linter().

paren_brace_linter() checks that there is a space between right parentheses and an opening curly brace. E.g., function(){} doesn't have a space, while function() {} does. Deprecated in favor of brace_linter().

semicolon_terminator_linter() checks that no semicolons terminate expressions. Deprecated in favor of semicolon_linter().

Usage

closed_curly_linter(allow_single_line = FALSE)

open_curly_linter(allow_single_line = FALSE)

paren_brace_linter()

semicolon_terminator_linter(semicolon = c("compound", "trailing"))

unneeded_concatenation_linter(allow_single_expression = TRUE)

single_quotes_linter()

consecutive_stopifnot_linter()

no_tab_linter()

Arguments

allow_single_line
if TRUE, allow an open and closed curly pair on the same line.

semicolon
A character vector defining which semicolons to report:

compound

Semicolons that separate two statements on the same line.

trailing

Semicolons following the last statement on the line.

Tags

configurable, deprecated, readability, style

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]