lintr-package {lintr}	R Documentation

Lintr

Description

Checks adherence to a given style, syntax errors, and possible semantic issues. Supports on the fly checking of R code edited with Emacs, Vim, and Sublime Text.

Author(s)

Maintainer: Michael Chirico michaelchirico4@gmail.com

Authors:

Jim Hester

Florent Angly (fangly)

Russ Hyde

Kun Ren

Alexander Rosenstock (AshesITR)

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets)

See Also

lint(), lint_package(), lint_dir(), linters

[Package lintr version 3.1.2 Index]