make_linter_from_xpath {lintr}	R Documentation

Create a linter from an XPath

Description

Create a linter from an XPath

Usage

make_linter_from_xpath(
  xpath,
  lint_message,
  type = c("warning", "style", "error"),
  level = c("expression", "file")
)

Arguments

xpath
Character string, an XPath identifying R code to lint. See xmlparsedata::xml_parse_data() and get_source_expressions().

lint_message
The message to be included as the message to the Lint object. If lint_message is a character vector the same length as xml, the i-th lint will be given the i-th message.

type
type of lint.

level
Which level of expression is being tested? "expression" means an individual expression, while "file" means all expressions in the current file are available.

Examples

number_linter <- make_linter_from_xpath("//NUM_CONST", "This is a number.")
lint(text = "1 + 2", linters = number_linter())

[Package lintr version 3.1.2 Index]