nonportable_path_linter {lintr}	R Documentation

Non-portable path linter

Description

Check that file.path() is used to construct safe and portable paths.

Usage

nonportable_path_linter(lax = TRUE)

Arguments

lax
Less stringent linting, leading to fewer false positives. If TRUE, only lint path strings, which

contain at least two path elements, with one having at least two characters and

contain only alphanumeric chars (including UTF-8), spaces, and win32-allowed punctuation

Tags

best_practices, configurable, robustness

See Also

linters for a complete list of linters available in lintr.

absolute_path_linter()

[Package lintr version 3.1.2 Index]