package_development_linters {lintr}	R Documentation

Package development linters

Description

Linters useful to package developers, for example for writing consistent tests.

Linters

The following linters are tagged with 'package_development':

backport_linter

conjunct_test_linter

expect_comparison_linter

expect_identical_linter

expect_length_linter

expect_named_linter

expect_not_linter

expect_null_linter

expect_s3_class_linter

expect_s4_class_linter

expect_true_false_linter

expect_type_linter

package_hooks_linter

yoda_test_linter

See Also

linters for a complete list of linters available in lintr.

[Package lintr version 3.1.2 Index]