parse_exclusions {lintr}	R Documentation

read a source file and parse all the excluded lines from it

Description

read a source file and parse all the excluded lines from it

Usage

parse_exclusions(file, exclude = settings$exclude,
  exclude_start = settings$exclude_start,
  exclude_end = settings$exclude_end)

Arguments

file
R source file

exclude
regular expression used to mark lines to exclude

exclude_start
regular expression used to mark the start of an excluded range

exclude_end
regular expression used to mark the end of an excluded range

[Package lintr version 1.0.3 Index]