read_settings {lintr}	R Documentation

Read lintr settings

Description

Lintr searches for settings for a given file in the following order.

options defined as linter.setting.

linter_file in the same directory

linter_file in the project directory

default_settings

Usage

read_settings(filename)

Arguments

filename
source file to be linted

Details

The default linter_file name is .lintr but it can be changed with option lintr.linter_file. This file is a dcf file, see read.dcf for details.

[Package lintr version 1.0.3 Index]