with_defaults {lintr}	R Documentation

Modify the list of default linters

Description

Modify the list of default linters

Usage

with_defaults(..., default = default_linters)

Arguments

...
named arguments of linters to change. If the named linter already exists it is replaced by the new linter, if it does not exist it is added. If the value is NULL the linter is removed.

default
default linters to change

Examples

# change the default line length cutoff
with_defaults(line_length_linter = line_length_linter(120))

# you can also omit the argument name if you are just using different
#   arguments.
with_defaults(line_length_linter(120))

# enforce camelCase rather than snake_case
with_defaults(camel_case_linter = NULL,
              snake_case_linter)

[Package lintr version 1.0.3 Index]