Create a linter configuration based on all available linters

Create a linter configuration based on all available linters

Usage

all_linters(packages = "lintr", ...)

Arguments

packages

A character vector of packages to search for linters.

...

Arguments of elements to change. If unnamed, the argument is automatically named. If the named argument already exists in the list of linters, it is replaced by the new element. If it does not exist, it is added. If the value is NULL, the linter is removed.

See also

Examples

names(all_linters())
#>  [1] "absolute_path_linter"             "any_duplicated_linter"           
#>  [3] "any_is_na_linter"                 "assignment_linter"               
#>  [5] "backport_linter"                  "boolean_arithmetic_linter"       
#>  [7] "brace_linter"                     "class_equals_linter"             
#>  [9] "commas_linter"                    "commented_code_linter"           
#> [11] "condition_message_linter"         "conjunct_test_linter"            
#> [13] "consecutive_assertion_linter"     "cyclocomp_linter"                
#> [15] "duplicate_argument_linter"        "empty_assignment_linter"         
#> [17] "equals_na_linter"                 "expect_comparison_linter"        
#> [19] "expect_identical_linter"          "expect_length_linter"            
#> [21] "expect_named_linter"              "expect_not_linter"               
#> [23] "expect_null_linter"               "expect_s3_class_linter"          
#> [25] "expect_s4_class_linter"           "expect_true_false_linter"        
#> [27] "expect_type_linter"               "extraction_operator_linter"      
#> [29] "fixed_regex_linter"               "for_loop_index_linter"           
#> [31] "function_argument_linter"         "function_left_parentheses_linter"
#> [33] "function_return_linter"           "if_not_else_linter"              
#> [35] "ifelse_censor_linter"             "implicit_assignment_linter"      
#> [37] "implicit_integer_linter"          "indentation_linter"              
#> [39] "infix_spaces_linter"              "inner_combine_linter"            
#> [41] "is_numeric_linter"                "keyword_quote_linter"            
#> [43] "length_levels_linter"             "lengths_linter"                  
#> [45] "library_call_linter"              "line_length_linter"              
#> [47] "literal_coercion_linter"          "matrix_apply_linter"             
#> [49] "missing_argument_linter"          "missing_package_linter"          
#> [51] "namespace_linter"                 "nested_ifelse_linter"            
#> [53] "nonportable_path_linter"          "numeric_leading_zero_linter"     
#> [55] "object_length_linter"             "object_name_linter"              
#> [57] "object_usage_linter"              "outer_negation_linter"           
#> [59] "package_hooks_linter"             "paren_body_linter"               
#> [61] "paste_linter"                     "pipe_call_linter"                
#> [63] "pipe_continuation_linter"         "quotes_linter"                   
#> [65] "redundant_equals_linter"          "redundant_ifelse_linter"         
#> [67] "regex_subset_linter"              "routine_registration_linter"     
#> [69] "semicolon_linter"                 "seq_linter"                      
#> [71] "sort_linter"                      "spaces_inside_linter"            
#> [73] "spaces_left_parentheses_linter"   "sprintf_linter"                  
#> [75] "string_boundary_linter"           "strings_as_factors_linter"       
#> [77] "system_file_linter"               "T_and_F_symbol_linter"           
#> [79] "todo_comment_linter"              "trailing_blank_lines_linter"     
#> [81] "trailing_whitespace_linter"       "undesirable_function_linter"     
#> [83] "undesirable_operator_linter"      "unnecessary_concatenation_linter"
#> [85] "unnecessary_lambda_linter"        "unnecessary_nested_if_linter"    
#> [87] "unnecessary_placeholder_linter"   "unreachable_code_linter"         
#> [89] "unused_import_linter"             "vector_logic_linter"             
#> [91] "whitespace_linter"                "yoda_test_linter"