Create a linter configuration based on all available linters

Create a linter configuration based on all available linters

Usage

all_linters(..., packages = "lintr")

Arguments

...

Arguments of elements to change. If unnamed, the argument is automatically named. If the named argument already exists in the list of linters, it is replaced by the new element. If it does not exist, it is added. If the value is NULL, the linter is removed.

packages

A character vector of packages to search for linters.

See also

Examples

names(all_linters())
#>   [1] "absolute_path_linter"             "any_duplicated_linter"           
#>   [3] "any_is_na_linter"                 "assignment_linter"               
#>   [5] "backport_linter"                  "boolean_arithmetic_linter"       
#>   [7] "brace_linter"                     "class_equals_linter"             
#>   [9] "coalesce_linter"                  "commas_linter"                   
#>  [11] "commented_code_linter"            "comparison_negation_linter"      
#>  [13] "condition_call_linter"            "condition_message_linter"        
#>  [15] "conjunct_test_linter"             "consecutive_assertion_linter"    
#>  [17] "consecutive_mutate_linter"        "cyclocomp_linter"                
#>  [19] "duplicate_argument_linter"        "empty_assignment_linter"         
#>  [21] "equals_na_linter"                 "expect_comparison_linter"        
#>  [23] "expect_identical_linter"          "expect_length_linter"            
#>  [25] "expect_named_linter"              "expect_not_linter"               
#>  [27] "expect_null_linter"               "expect_s3_class_linter"          
#>  [29] "expect_s4_class_linter"           "expect_true_false_linter"        
#>  [31] "expect_type_linter"               "fixed_regex_linter"              
#>  [33] "for_loop_index_linter"            "function_argument_linter"        
#>  [35] "function_left_parentheses_linter" "function_return_linter"          
#>  [37] "if_not_else_linter"               "if_switch_linter"                
#>  [39] "ifelse_censor_linter"             "implicit_assignment_linter"      
#>  [41] "implicit_integer_linter"          "indentation_linter"              
#>  [43] "infix_spaces_linter"              "inner_combine_linter"            
#>  [45] "is_numeric_linter"                "keyword_quote_linter"            
#>  [47] "length_levels_linter"             "length_test_linter"              
#>  [49] "lengths_linter"                   "library_call_linter"             
#>  [51] "line_length_linter"               "list_comparison_linter"          
#>  [53] "literal_coercion_linter"          "matrix_apply_linter"             
#>  [55] "missing_argument_linter"          "missing_package_linter"          
#>  [57] "namespace_linter"                 "nested_ifelse_linter"            
#>  [59] "nested_pipe_linter"               "nonportable_path_linter"         
#>  [61] "nrow_subset_linter"               "numeric_leading_zero_linter"     
#>  [63] "nzchar_linter"                    "object_length_linter"            
#>  [65] "object_name_linter"               "object_overwrite_linter"         
#>  [67] "object_usage_linter"              "one_call_pipe_linter"            
#>  [69] "outer_negation_linter"            "package_hooks_linter"            
#>  [71] "paren_body_linter"                "paste_linter"                    
#>  [73] "pipe_call_linter"                 "pipe_consistency_linter"         
#>  [75] "pipe_continuation_linter"         "pipe_return_linter"              
#>  [77] "print_linter"                     "quotes_linter"                   
#>  [79] "redundant_equals_linter"          "redundant_ifelse_linter"         
#>  [81] "regex_subset_linter"              "rep_len_linter"                  
#>  [83] "repeat_linter"                    "return_linter"                   
#>  [85] "routine_registration_linter"      "sample_int_linter"               
#>  [87] "scalar_in_linter"                 "semicolon_linter"                
#>  [89] "seq_linter"                       "sort_linter"                     
#>  [91] "spaces_inside_linter"             "spaces_left_parentheses_linter"  
#>  [93] "sprintf_linter"                   "stopifnot_all_linter"            
#>  [95] "string_boundary_linter"           "strings_as_factors_linter"       
#>  [97] "system_file_linter"               "T_and_F_symbol_linter"           
#>  [99] "terminal_close_linter"            "todo_comment_linter"             
#> [101] "trailing_blank_lines_linter"      "trailing_whitespace_linter"      
#> [103] "undesirable_function_linter"      "undesirable_operator_linter"     
#> [105] "unnecessary_concatenation_linter" "unnecessary_lambda_linter"       
#> [107] "unnecessary_nesting_linter"       "unnecessary_placeholder_linter"  
#> [109] "unreachable_code_linter"          "unused_import_linter"            
#> [111] "vector_logic_linter"              "which_grepl_linter"              
#> [113] "whitespace_linter"                "yoda_test_linter"