Create a linter configuration based on all available linters

Create a linter configuration based on all available linters

Usage

all_linters(..., packages = "lintr")

Arguments

...

Arguments of elements to change. If unnamed, the argument is automatically named. If the named argument already exists in the list of linters, it is replaced by the new element. If it does not exist, it is added. If the value is NULL, the linter is removed.

packages

A character vector of packages to search for linters.

See also

Examples

names(all_linters())
#>   [1] "absolute_path_linter"             "any_duplicated_linter"           
#>   [3] "any_is_na_linter"                 "assignment_linter"               
#>   [5] "backport_linter"                  "boolean_arithmetic_linter"       
#>   [7] "brace_linter"                     "class_equals_linter"             
#>   [9] "commas_linter"                    "commented_code_linter"           
#>  [11] "comparison_negation_linter"       "condition_call_linter"           
#>  [13] "condition_message_linter"         "conjunct_test_linter"            
#>  [15] "consecutive_assertion_linter"     "consecutive_mutate_linter"       
#>  [17] "cyclocomp_linter"                 "duplicate_argument_linter"       
#>  [19] "empty_assignment_linter"          "equals_na_linter"                
#>  [21] "expect_comparison_linter"         "expect_identical_linter"         
#>  [23] "expect_length_linter"             "expect_named_linter"             
#>  [25] "expect_not_linter"                "expect_null_linter"              
#>  [27] "expect_s3_class_linter"           "expect_s4_class_linter"          
#>  [29] "expect_true_false_linter"         "expect_type_linter"              
#>  [31] "fixed_regex_linter"               "for_loop_index_linter"           
#>  [33] "function_argument_linter"         "function_left_parentheses_linter"
#>  [35] "function_return_linter"           "if_not_else_linter"              
#>  [37] "if_switch_linter"                 "ifelse_censor_linter"            
#>  [39] "implicit_assignment_linter"       "implicit_integer_linter"         
#>  [41] "indentation_linter"               "infix_spaces_linter"             
#>  [43] "inner_combine_linter"             "is_numeric_linter"               
#>  [45] "keyword_quote_linter"             "length_levels_linter"            
#>  [47] "length_test_linter"               "lengths_linter"                  
#>  [49] "library_call_linter"              "line_length_linter"              
#>  [51] "list_comparison_linter"           "literal_coercion_linter"         
#>  [53] "matrix_apply_linter"              "missing_argument_linter"         
#>  [55] "missing_package_linter"           "namespace_linter"                
#>  [57] "nested_ifelse_linter"             "nested_pipe_linter"              
#>  [59] "nonportable_path_linter"          "nrow_subset_linter"              
#>  [61] "numeric_leading_zero_linter"      "nzchar_linter"                   
#>  [63] "object_length_linter"             "object_name_linter"              
#>  [65] "object_overwrite_linter"          "object_usage_linter"             
#>  [67] "one_call_pipe_linter"             "outer_negation_linter"           
#>  [69] "package_hooks_linter"             "paren_body_linter"               
#>  [71] "paste_linter"                     "pipe_call_linter"                
#>  [73] "pipe_consistency_linter"          "pipe_continuation_linter"        
#>  [75] "pipe_return_linter"               "print_linter"                    
#>  [77] "quotes_linter"                    "redundant_equals_linter"         
#>  [79] "redundant_ifelse_linter"          "regex_subset_linter"             
#>  [81] "rep_len_linter"                   "repeat_linter"                   
#>  [83] "return_linter"                    "routine_registration_linter"     
#>  [85] "sample_int_linter"                "scalar_in_linter"                
#>  [87] "semicolon_linter"                 "seq_linter"                      
#>  [89] "sort_linter"                      "spaces_inside_linter"            
#>  [91] "spaces_left_parentheses_linter"   "sprintf_linter"                  
#>  [93] "stopifnot_all_linter"             "string_boundary_linter"          
#>  [95] "strings_as_factors_linter"        "system_file_linter"              
#>  [97] "T_and_F_symbol_linter"            "terminal_close_linter"           
#>  [99] "todo_comment_linter"              "trailing_blank_lines_linter"     
#> [101] "trailing_whitespace_linter"       "undesirable_function_linter"     
#> [103] "undesirable_operator_linter"      "unnecessary_concatenation_linter"
#> [105] "unnecessary_lambda_linter"        "unnecessary_nesting_linter"      
#> [107] "unnecessary_placeholder_linter"   "unreachable_code_linter"         
#> [109] "unused_import_linter"             "vector_logic_linter"             
#> [111] "which_grepl_linter"               "whitespace_linter"               
#> [113] "yoda_test_linter"