expect_lint_free {lintr}	R Documentation

Test that the package is lint free

Description

This function is a thin wrapper around lint_package that simply tests there are no lints in the package. It can be used to ensure that your tests fail if the package contains lints.

Usage

expect_lint_free(...)

Arguments

...
arguments passed to lint_package()

[Package lintr version 3.1.2 Index]