Library call linter

This linter covers several rules related to library() calls:

Usage

library_call_linter(allow_preamble = TRUE)

Arguments

allow_preamble

Logical, default TRUE. If FALSE, no code is allowed to precede the first library() call, otherwise some setup code is allowed, but all library() calls must follow consecutively after the first one.

Details

Enforce such calls to all be at the top of the script.

Block usage of argument character.only, in particular for loading packages in a loop.

Block consecutive calls to suppressMessages(library(.)) in favor of using suppressMessages() only once to suppress messages from all library() calls. Ditto suppressPackageStartupMessages().

See also

Tags

best_practices, configurable, readability, style

Examples

# will produce lints

code <- "library(dplyr)\nprint('test')\nlibrary(tidyr)"
writeLines(code)
#> library(dplyr)
#> print('test')
#> library(tidyr)
lint(
  text = code,
  linters = library_call_linter()
)
#> ::warning file=<text>,line=3,col=1::file=<text>,line=3,col=1,[library_call_linter] Move all library calls to the top of the script.

lint(
  text = "library('dplyr', character.only = TRUE)",
  linters = library_call_linter()
)
#> ::warning file=<text>,line=1,col=1::file=<text>,line=1,col=1,[library_call_linter] Use symbols in library calls to avoid the need for 'character.only'.

code <- paste(
  "pkg <- c('dplyr', 'tibble')",
  "sapply(pkg, library, character.only = TRUE)",
  sep = "\n"
)
writeLines(code)
#> pkg <- c('dplyr', 'tibble')
#> sapply(pkg, library, character.only = TRUE)
lint(
  text = code,
  linters = library_call_linter()
)
#> ::warning file=<text>,line=2,col=1::file=<text>,line=2,col=1,[library_call_linter] Call library() directly, not vectorized with sapply().

code <- "suppressMessages(library(dplyr))\nsuppressMessages(library(tidyr))"
writeLines(code)
#> suppressMessages(library(dplyr))
#> suppressMessages(library(tidyr))
lint(
  text = code,
  linters = library_call_linter()
)
#> ::warning file=<text>,line=1,col=1::file=<text>,line=1,col=1,[library_call_linter] Unify consecutive calls to suppressMessages(). You can do so by writing all of the calls in one braced expression like suppressMessages({...}).

# okay
code <- "library(dplyr)\nprint('test')"
writeLines(code)
#> library(dplyr)
#> print('test')
lint(
  text = code,
  linters = library_call_linter()
)

code <- "# comment\nlibrary(dplyr)"
lint(
  text = code,
  linters = library_call_linter()
)

code <- paste(
  "foo <- function(pkg) {",
  "  sapply(pkg, library, character.only = TRUE)",
  "}",
  sep = "\n"
)
writeLines(code)
#> foo <- function(pkg) {
#>   sapply(pkg, library, character.only = TRUE)
#> }
lint(
  text = code,
  linters = library_call_linter()
)

code <- "suppressMessages({\n  library(dplyr)\n  library(tidyr)\n})"
writeLines(code)
#> suppressMessages({
#>   library(dplyr)
#>   library(tidyr)
#> })
lint(
  text = code,
  linters = library_call_linter()
)