Available linters

A variety of linters is available in lintr. The most popular ones are readily accessible through default_linters().

Within a lint() function call, the linters in use are initialized with the provided arguments and fed with the source file (provided by get_source_expressions()).

A data frame of all available linters can be retrieved using available_linters(). Documentation for linters is structured into tags to allow for easier discovery; see also available_tags().

Tags

The following tags exist:

best_practices (36 linters)

common_mistakes (6 linters)

configurable (20 linters)

consistency (16 linters)

correctness (7 linters)

default (24 linters)

deprecated (4 linters)

efficiency (16 linters)

executing (5 linters)

package_development (14 linters)

readability (37 linters)

robustness (12 linters)

style (35 linters)

Linters

The following linters exist:

absolute_path_linter (tags: best_practices, configurable, robustness)

any_duplicated_linter (tags: best_practices, efficiency)

any_is_na_linter (tags: best_practices, efficiency)

assignment_linter (tags: consistency, default, style)

backport_linter (tags: configurable, package_development, robustness)

brace_linter (tags: configurable, default, readability, style)

class_equals_linter (tags: best_practices, consistency, robustness)

closed_curly_linter (tags: configurable, deprecated, readability, style)

commas_linter (tags: default, readability, style)

commented_code_linter (tags: best_practices, default, readability, style)

condition_message_linter (tags: best_practices, consistency)

conjunct_test_linter (tags: best_practices, package_development, readability)

consecutive_stopifnot_linter (tags: consistency, readability, style)

cyclocomp_linter (tags: best_practices, configurable, default, readability, style)

duplicate_argument_linter (tags: common_mistakes, configurable, correctness)

equals_na_linter (tags: common_mistakes, correctness, default, robustness)

expect_comparison_linter (tags: best_practices, package_development)

expect_identical_linter (tags: package_development)

expect_length_linter (tags: best_practices, package_development, readability)

expect_named_linter (tags: best_practices, package_development, readability)

expect_not_linter (tags: best_practices, package_development, readability)

expect_null_linter (tags: best_practices, package_development)

expect_s3_class_linter (tags: best_practices, package_development)

expect_s4_class_linter (tags: best_practices, package_development)

expect_true_false_linter (tags: best_practices, package_development, readability)

expect_type_linter (tags: best_practices, package_development)

extraction_operator_linter (tags: best_practices, style)

fixed_regex_linter (tags: best_practices, efficiency, readability)

function_left_parentheses_linter (tags: default, readability, style)

ifelse_censor_linter (tags: best_practices, efficiency)

implicit_integer_linter (tags: best_practices, consistency, style)

infix_spaces_linter (tags: default, readability, style)

inner_combine_linter (tags: consistency, efficiency, readability)

line_length_linter (tags: configurable, default, readability, style)

literal_coercion_linter (tags: best_practices, consistency, efficiency)

missing_argument_linter (tags: common_mistakes, configurable, correctness)

missing_package_linter (tags: common_mistakes, robustness)

namespace_linter (tags: configurable, correctness, executing, robustness)

nested_ifelse_linter (tags: efficiency, readability)

no_tab_linter (tags: consistency, default, style)

nonportable_path_linter (tags: best_practices, configurable, robustness)

numeric_leading_zero_linter (tags: consistency, readability, style)

object_length_linter (tags: configurable, default, executing, readability, style)

object_name_linter (tags: configurable, consistency, default, executing, style)

object_usage_linter (tags: correctness, default, executing, readability, style)

open_curly_linter (tags: configurable, deprecated, readability, style)

outer_negation_linter (tags: best_practices, efficiency, readability)

package_hooks_linter (tags: correctness, package_development, style)

paren_body_linter (tags: default, readability, style)

paren_brace_linter (tags: deprecated, readability, style)

paste_linter (tags: best_practices, consistency)

pipe_call_linter (tags: readability, style)

pipe_continuation_linter (tags: default, readability, style)

redundant_ifelse_linter (tags: best_practices, consistency, efficiency)

regex_subset_linter (tags: best_practices, efficiency)

semicolon_linter (tags: configurable, default, readability, style)

semicolon_terminator_linter (tags: configurable, deprecated, readability, style)

seq_linter (tags: best_practices, consistency, default, efficiency, robustness)

single_quotes_linter (tags: consistency, default, readability, style)

spaces_inside_linter (tags: default, readability, style)

spaces_left_parentheses_linter (tags: default, readability, style)

sprintf_linter (tags: common_mistakes, correctness)

string_boundary_linter (tags: efficiency, readability)

strings_as_factors_linter (tags: robustness)

system_file_linter (tags: best_practices, consistency, readability)

T_and_F_symbol_linter (tags: best_practices, consistency, default, readability, robustness, style)

todo_comment_linter (tags: configurable, style)

trailing_blank_lines_linter (tags: default, style)

trailing_whitespace_linter (tags: default, style)

undesirable_function_linter (tags: best_practices, configurable, efficiency, robustness, style)

undesirable_operator_linter (tags: best_practices, configurable, efficiency, robustness, style)

unneeded_concatenation_linter (tags: configurable, efficiency, readability, style)

unreachable_code_linter (tags: best_practices, readability)

unused_import_linter (tags: best_practices, common_mistakes, configurable, executing)

vector_logic_linter (tags: best_practices, default, efficiency)

yoda_test_linter (tags: best_practices, package_development, readability)